Concordance text by choice of tag

This function produces concordance lines of text from data by finding up to two tag matches in tokenized text.

Usage

conc_by_tag(
  data,
  what = "token",
  tag = "mda",
  match,
  cols = c("corpus", "doc_id", "sentence"),
  tag2 = NULL,
  match2 = NULL,
  ...
)

Arguments

data

A relational data frame containing the text to concordance. The data frame is expected to have one column of tokens, in tokenized form, at least one column of the corresponding tags, and identifying details such as corpus, doc_id etc.

what

The name of the column containing the text to concatenate. Default is "token".

tag

The name of the column containing the tags to match. Default is "mda_tags".

match

The tag to match within the tag column.

cols

The names of the columns to include in the output. By default, these will include corpus, doc_id, sentence, but need to be changed if the inputting df has different columns.

tag2

The name of the second column containing the tags to match (optional).

match2

The second tag to match within the tag2 column (optional).

...

Additional arguments to be passed onto dtagger::quick_conc.

Value

A data frame containing the cols plus an extra column with the concatenated text.

Details

The purpose of this function is to allow fine-grained concordance searches of tagged text. The input should be a dataframe with a column for tokens in tokenized form, and separate columns for tags, document and corpus details.

Typically, the function can be used with output from udpipe::udpipe_annotate and dtagger::dtag_tbl od dtagger::dtag_directory functions.

The concordancer can take up to two tag inputs, for example matching all upos == "ADJ" tags and dep_rel == "amod" tags, and seing the resulting key words in context.